s -> s*100

+++ instruments/gasoanalisator/gashandler.py
@@ -72,9 +72,8 @@ class GasoanalysatorHandler(Instrument):
         """
         answer = {}
         for i, paramname in enumerate(["CO", "CH", "CO2", "O2", "NO", "L"]):
-            answer[paramname] = int(
-                float_from_bytes(content[6 * i : 6 * (i + 1)]) * 10**4
-            )
+            answer[paramname] = float_from_bytes(content[6 * i : 6 * (i + 1)]) * 10**4
+            
 
         return self.make_answer(GIApiStatus.data, content=answer)
